+++ src/valentina/cogs/experience.py
@@ -6,8 +6,8 @@ from discord.commands import Option
 from discord.ext import commands
 
 from valentina.constants import TraitCategory, XPMultiplier
+from valentina.models import User
 from valentina.models.bot import Valentina, ValentinaContext
-from valentina.models.mongo_collections import User
 from valentina.utils.converters import (
     ValidCharacterObject,
     ValidCharTrait,
