+++ src/valentina/cogs/macros.py
@@ -5,8 +5,8 @@ import discord
 from discord.commands import Option
 from discord.ext import commands
 
+from valentina.models import User, UserMacro
 from valentina.models.bot import Valentina, ValentinaContext
-from valentina.models.mongo_collections import User, UserMacro
 from valentina.utils.helpers import truncate_string
 from valentina.utils.options import (
     select_char_trait,
