+++ src/valentina/models/__init__.py
@@ -1,9 +1,33 @@
 """Models for Valentina."""
 
-from .probability import Probability
-from .statistics import Statistics
+from .campaign import Campaign, CampaignChapter, CampaignExperience, CampaignNote, CampaignNPC
+from .character import Character, CharacterSheetSection, CharacterTrait
+from .database import GlobalProperty
+from .guild import Guild, GuildChannels, GuildPermissions, GuildRollResultThumbnail
+from .user import User, UserMacro
+
+from .statistics import Statistics, RollStatistic  # isort: skip
+from .probability import Probability, RollProbability  # isort: skip
+
 
 __all__ = [
+    "Campaign",
+    "CampaignChapter",
+    "CampaignExperience",
+    "CampaignNote",
+    "CampaignNPC",
+    "Character",
+    "CharacterTrait",
+    "GlobalProperty",
+    "Guild",
+    "GuildChannels",
+    "GuildPermissions",
+    "GuildRollResultThumbnail",
+    "CharacterSheetSection",
     "Probability",
+    "RollProbability",
+    "RollStatistic",
     "Statistics",
+    "User",
+    "UserMacro",
 ]
