+++ src/valentina/utils/perform_roll.py
@@ -1,8 +1,8 @@
 """Perform a diceroll."""
 
+from valentina.models import Character, CharacterTrait
 from valentina.models.bot import ValentinaContext
 from valentina.models.dicerolls import DiceRoll
-from valentina.models.mongo_collections import Character, CharacterTrait
 from valentina.views import ReRollButton, RollDisplay
